Home

+++ src/main/java/com/javaonlinecourse/b2lesson2/homework/hw01/HomeTask02.java
@@ -6,9 +6,9 @@ package com.javaonlinecourse.b2lesson2.homework.hw01;
  */
 public class HomeTask02 {
     public static void main(String[] args) {
-        int alpha = 15;
-        int beta = 4;
-        float gamma = alpha / beta;
+        double alpha = 15;
+        double beta = 4;
+        double gamma = alpha / beta;
         double delta = alpha * 1e-3 + gamma;
         System.out.println(delta);
     }
